requirements.txt
python-telegram-bot==13.15
Flask==2.2.5
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
